[API-45] Mappers segregated by types
