+++ daos/channels.py
@@ -1,12 +1,10 @@
 from app import db
 
 from daos.database import DatabaseClient
-from daos.builder import TableEntryBuilder, ModelBuilder
+from daos.mappers.channels import ChannelDatabaseMapper, ChannelModelMapper
 
-from tables.messages import *
-from tables.channels import *
-from tables.users import *
-from tables.teams import *
+from tables.channels import ChannelTableEntry
+from tables.users import UserTableEntry, UsersByChannelsTableEntry
 
 from sqlalchemy import and_
 
@@ -15,12 +13,12 @@ class ChannelDatabaseClient:
 
     @classmethod
     def add_channel(cls, channel):
-        channel_entry = TableEntryBuilder.to_channel(channel)
+        channel_entry = ChannelDatabaseMapper.to_channel(channel)
         DatabaseClient.add(channel_entry)
 
     @classmethod
     def add_channel_user(cls, channel_user):
-        channel_user_entry = TableEntryBuilder.to_channel_user(channel_user)
+        channel_user_entry = ChannelDatabaseMapper.to_channel_user(channel_user)
         DatabaseClient.add(channel_user_entry)
 
     @classmethod
@@ -49,14 +47,14 @@ class ChannelDatabaseClient:
         channel_entry = db.session.query(ChannelTableEntry).filter(
             ChannelTableEntry.channel_id == channel_id
         ).one_or_none()
-        return ModelBuilder.to_channel(channel_entry)
+        return ChannelModelMapper.to_channel(channel_entry)
 
     @classmethod
     def get_channel_by_name(cls, channel_name):
         channel_entry = db.session.query(ChannelTableEntry).filter(
             ChannelTableEntry.name == channel_name
         ).one_or_none()
-        return ModelBuilder.to_channel(channel_entry)
+        return ChannelModelMapper.to_channel(channel_entry)
 
     @classmethod
     def get_user_in_channel_by_ids(cls, user_id, channel_id):
@@ -64,7 +62,7 @@ class ChannelDatabaseClient:
             UsersByChannelsTableEntry.user_id == user_id,
             UsersByChannelsTableEntry.channel_id == channel_id
         ).one_or_none()
-        return ModelBuilder.to_user_in_channel(channel_user_entry)
+        return ChannelModelMapper.to_user_in_channel(channel_user_entry)
 
     @classmethod
     def get_all_channel_users_by_channel_id(cls, channel_id, ignored_user_id=None):
@@ -85,4 +83,4 @@ class ChannelDatabaseClient:
                 UsersByChannelsTableEntry.user_id != ignored_user_id
             )
         ).all()
-        return ModelBuilder.to_public_users(users)
+        return ChannelModelMapper.to_channel_members(users)
